Leveling System and SawLauncher
Implement Leveling System, Saw Launcher weapon, projectile scripts future proofed to handle damaging enemy.

+++ Assets/Scripts/Weapons/Weapon.cs
@@ -14,7 +14,7 @@ public abstract class Weapon : MonoBehaviour
     public float baseDespawnTime;
     
     // Updated Variables depending on weapon level.
-    public float damage => baseDamage * level;
+    public int damage => baseDamage * level;
     public float fireRate => baseFireRate;
     public float projectileSpeed => baseProjectileSpeed + level * 1.2f;
     public float despawnTime => baseDespawnTime;
